[BACKLOG-5396] Rename integration tests in Kettle engine module
