+++ engine/test-src/org/pentaho/di/trans/steps/rest/RestIT.java
@@ -20,7 +20,7 @@
  *
  ******************************************************************************/
 
-package org.pentaho.di.trans.steps.http;
+package org.pentaho.di.trans.steps.rest;
 
 import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.mock;
@@ -29,8 +29,6 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.pentaho.di.core.util.Assert.assertTrue;
 
-import java.io.IOException;
-import java.net.InetSocketAddress;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
@@ -40,7 +38,6 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
-import org.mockito.InjectMocks;
 import org.pentaho.di.core.KettleClientEnvironment;
 import org.pentaho.di.core.exception.KettleException;
 import org.pentaho.di.core.exception.KettleStepException;
@@ -56,23 +53,29 @@ import org.pentaho.di.trans.steps.loadsave.validator.FieldLoadSaveValidator;
 import org.pentaho.di.trans.steps.loadsave.validator.StringLoadSaveValidator;
 import org.pentaho.di.trans.steps.mock.StepMockHelper;
 
-import com.sun.net.httpserver.HttpExchange;
-import com.sun.net.httpserver.HttpHandler;
+import com.sun.jersey.api.container.httpserver.HttpServerFactory;
 import com.sun.net.httpserver.HttpServer;
 
 /**
- * User: Dzmitry Stsiapanau Date: 12/2/13 Time: 1:24 PM
+ * User: Dzmitry Stsiapanau Date: 11/29/13 Time: 3:42 PM
  */
-public class HTTPTest {
+public class RestIT {
 
-  private class HTTPHandler extends HTTP {
+  public static final String HTTP_LOCALHOST_9998 = "http://localhost:9998/";
+
+  private class RestHandler extends Rest {
 
     Object[] row = new Object[] { "anyData" };
     Object[] outputRow;
 
-    public HTTPHandler( StepMeta stepMeta, StepDataInterface stepDataInterface, int copyNr, TransMeta transMeta,
+    public RestHandler( StepMeta stepMeta, StepDataInterface stepDataInterface, int copyNr, TransMeta transMeta,
         Trans trans ) {
       super( stepMeta, stepDataInterface, copyNr, transMeta, trans );
+    }
+
+    @SuppressWarnings( "unused" )
+    public void setRow( Object[] row ) {
+      this.row = row;
     }
 
     /**
@@ -105,12 +108,8 @@ public class HTTPTest {
 
   }
 
-  public static final String host = "localhost";
-  public static final int port = 9998;
-  public static final String HTTP_LOCALHOST_9998 = "http://localhost:9998/";
-  @InjectMocks
-  private StepMockHelper<HTTPMeta, HTTPData> stepMockHelper;
-  private HttpServer httpServer;
+  private StepMockHelper<RestMeta, RestData> stepMockHelper;
+  private HttpServer server;
 
   @BeforeClass
   public static void setupBeforeClass() throws KettleException {
@@ -119,76 +118,65 @@ public class HTTPTest {
 
   @Before
   public void setUp() throws Exception {
-    stepMockHelper = new StepMockHelper<HTTPMeta, HTTPData>( "HTTP CLIENT TEST", HTTPMeta.class, HTTPData.class );
+    stepMockHelper = new StepMockHelper<RestMeta, RestData>( "REST CLIENT TEST", RestMeta.class, RestData.class );
     when( stepMockHelper.logChannelInterfaceFactory.create( any(), any( LoggingObjectInterface.class ) ) ).thenReturn(
         stepMockHelper.logChannelInterface );
     when( stepMockHelper.trans.isRunning() ).thenReturn( true );
     verify( stepMockHelper.trans, never() ).stopAll();
-    startHttp204Answer();
+    server = HttpServerFactory.create( HTTP_LOCALHOST_9998 );
+    server.start();
   }
 
   @After
   public void tearDown() throws Exception {
-    httpServer.stop( 5 );
-
+    server.stop( 0 );
   }
 
   @Test
-  public void test204Answer() throws Exception {
-    HTTPData data = new HTTPData();
+  public void testNoContent() throws Exception {
+    RestData data = new RestData();
     Object[] expectedRow = new Object[] { "", 204L, null, null, null, null, null, null, null, null, null, null };
-    HTTP http = new HTTPHandler( stepMockHelper.stepMeta, data, 0, stepMockHelper.transMeta, stepMockHelper.trans );
+    Rest rest = new RestHandler( stepMockHelper.stepMeta, data, 0, stepMockHelper.transMeta, stepMockHelper.trans );
     RowMetaInterface inputRowMeta = mock( RowMetaInterface.class );
-    http.setInputRowMeta( inputRowMeta );
+    rest.setInputRowMeta( inputRowMeta );
     when( inputRowMeta.clone() ).thenReturn( inputRowMeta );
-    when( stepMockHelper.processRowsStepMetaInterface.getUrl() ).thenReturn( HTTP_LOCALHOST_9998 );
-    when( stepMockHelper.processRowsStepMetaInterface.getHeaderField() ).thenReturn( new String[] {} );
-    when( stepMockHelper.processRowsStepMetaInterface.getArgumentField() ).thenReturn( new String[] {} );
-    when( stepMockHelper.processRowsStepMetaInterface.getResultCodeFieldName() ).thenReturn( "ResultCodeFieldName" );
-    when( stepMockHelper.processRowsStepMetaInterface.getFieldName() ).thenReturn( "ResultFieldName" );
-    http.init( stepMockHelper.processRowsStepMetaInterface, data );
-    assertTrue( http.processRow( stepMockHelper.processRowsStepMetaInterface, data ) );
+    when( stepMockHelper.processRowsStepMetaInterface.getUrl() ).thenReturn(
+        HTTP_LOCALHOST_9998 + "restTest/restNoContentAnswer" );
+    when( stepMockHelper.processRowsStepMetaInterface.getMethod() ).thenReturn( RestMeta.HTTP_METHOD_GET );
+    rest.init( stepMockHelper.processRowsStepMetaInterface, data );
+    data.resultFieldName = "ResultFieldName";
+    data.resultCodeFieldName = "ResultCodeFieldName";
+    assertTrue( rest.processRow( stepMockHelper.processRowsStepMetaInterface, data ) );
     System.out.println( Arrays.toString( expectedRow ) );
-    Object[] out = ( (HTTPHandler) http ).getOutputRow();
+    Object[] out = ( (RestHandler) rest ).getOutputRow();
     System.out.println( Arrays.toString( out ) );
     assertTrue( Arrays.equals( expectedRow, out ) );
-  }
-
-  private void startHttp204Answer() throws IOException {
-    httpServer = HttpServer.create( new InetSocketAddress( HTTPTest.host, HTTPTest.port ), 10 );
-    httpServer.createContext( "/", new HttpHandler() {
-      @Override
-      public void handle( HttpExchange httpExchange ) throws IOException {
-        httpExchange.sendResponseHeaders( 204, 0 );
-        httpExchange.close();
-      }
-    } );
-    httpServer.start();
   }
 
   @Test
   public void testLoadSaveRoundTrip() throws KettleException {
     List<String> attributes =
-        Arrays.asList( "url", "urlInField", "urlField", "encoding", "httpLogin", "httpPassword", "proxyHost",
-            "proxyPort", "socketTimeout", "connectionTimeout", "closeIdleConnectionsTime", "argumentField",
-            "argumentParameter", "headerField", "headerParameter", "fieldName", "resultCodeFieldName",
-            "responseTimeFieldName" );
+        Arrays.asList( "applicationType", "method", "url", "urlInField", "dynamicMethod", "methodFieldName",
+            "urlField", "bodyField", "httpLogin", "httpPassword", "proxyHost", "proxyPort", "preemptive",
+            "trustStoreFile", "trustStorePassword", "headerField", "headerName", "parameterField", "parameterName",
+            "matrixParameterField", "matrixParameterName", "fieldName", "resultCodeFieldName", "responseTimeFieldName" );
 
     Map<String, FieldLoadSaveValidator<?>> fieldLoadSaveValidatorAttributeMap =
         new HashMap<String, FieldLoadSaveValidator<?>>();
 
-    //Arrays need to be consistent length
+    // Arrays need to be consistent length
     FieldLoadSaveValidator<String[]> stringArrayLoadSaveValidator =
         new ArrayLoadSaveValidator<String>( new StringLoadSaveValidator(), 25 );
-    fieldLoadSaveValidatorAttributeMap.put( "argumentField", stringArrayLoadSaveValidator );
-    fieldLoadSaveValidatorAttributeMap.put( "argumentParameter", stringArrayLoadSaveValidator );
     fieldLoadSaveValidatorAttributeMap.put( "headerField", stringArrayLoadSaveValidator );
-    fieldLoadSaveValidatorAttributeMap.put( "headerParameter", stringArrayLoadSaveValidator );
+    fieldLoadSaveValidatorAttributeMap.put( "headerName", stringArrayLoadSaveValidator );
+    fieldLoadSaveValidatorAttributeMap.put( "parameterField", stringArrayLoadSaveValidator );
+    fieldLoadSaveValidatorAttributeMap.put( "parameterName", stringArrayLoadSaveValidator );
+    fieldLoadSaveValidatorAttributeMap.put( "matrixParameterField", stringArrayLoadSaveValidator );
+    fieldLoadSaveValidatorAttributeMap.put( "matrixParameterName", stringArrayLoadSaveValidator );
 
     LoadSaveTester loadSaveTester =
-        new LoadSaveTester( HTTPMeta.class, attributes, new HashMap<String, String>(),
-            new HashMap<String, String>(), fieldLoadSaveValidatorAttributeMap,
-            new HashMap<String, FieldLoadSaveValidator<?>>() );
+        new LoadSaveTester( RestMeta.class, attributes, new HashMap<String, String>(), new HashMap<String, String>(),
+            fieldLoadSaveValidatorAttributeMap, new HashMap<String, FieldLoadSaveValidator<?>>() );
 
     loadSaveTester.testRepoRoundTrip();
     loadSaveTester.testXmlRoundTrip();
